# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Praefectus Daemon API",
    "description": "REST API for the Praefectus AI session orchestrator daemon.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/vigil/chat": {
      "post": {
        "operationId": "vigilChat",
        "summary": "Chat with the global Vigil overseer",
        "tags": ["Vigil"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vigil chat response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "type": "string" }
                  },
                  "required": ["response"]
                }
              }
            }
          }
        }
      }
    },
    "/api/vigil/history": {
      "get": {
        "operationId": "getVigilHistory",
        "summary": "Get Vigil chat history",
        "tags": ["Vigil"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 100 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat message history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["messages"],
                  "properties": {
                    "messages": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/VigilMessage" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vigil/acta": {
      "get": {
        "operationId": "getActa",
        "summary": "Get the acta (briefing) for a project",
        "tags": ["Vigil"],
        "parameters": [
          {
            "name": "projectPath",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Project acta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projectPath": { "type": "string" },
                    "acta": {
                      "type": ["string", "null"],
                      "description": "Null if no acta exists for the project"
                    }
                  },
                  "required": ["projectPath", "acta"]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatInput": {
        "type": "object",
        "properties": {
          "projectPath": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["message"]
      },
      "VigilMessage": {
        "type": "object",
        "required": ["id", "role", "content", "createdAt"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "role": { "type": "string", "enum": ["user", "vigil"] },
          "content": { "type": "string" },
          "embeddedCards": { "type": ["string", "null"] },
          "createdAt": { "type": "integer", "format": "int64" }
        }
      },
...
    }
  },
...
}